feat(nextbike): add cli commands
